% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageName.R
\docType{data}
\name{dataset_name}
\alias{dataset_name}
\title{Title of built-in dataset}
\format{
A \code{data.frame}/\code{list}/... with ... .
}
\source{
\href{link}{link}. Retrieved month day, year.
}
\usage{
data("dataset_name")
}
\description{
Description of data.
}
\examples{
data("dataset_name", package = "packageName")
dataset_name[...]

}
\keyword{datasets}
