% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageName.R
\docType{package}
\name{packageName-package}
\alias{packageName}
\alias{packageName-package}
\title{packageName: What your package does}
\description{
The \pkg{packageName} package is fantastic at doing x and y.
}
\section{Main functionality}{

\itemize{
\item Functionality A: \code{\link{fct_x}}, \code{\link{fct_y}}
\item Functionality B: \code{\link{fct_e}}, \code{\link{fct_f}}
}
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
