% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\alias{release.specimens}
\alias{release,specimens-method}
\title{Release a table for an herbarium}
\usage{
release(x, ...)

\method{release}{specimens}(x, herb, ...)
}
\arguments{
\item{x}{A \linkS4class{specimens} object retrieved by \code{\link[=read_spec]{read_spec()}}.}

\item{...}{Further arguments passed among methods (not in use).}

\item{herb}{A character value containig the code of the herbarium for the
release.}
}
\value{
A data frame with columns sorted and named for release at a respective
herbarium.
}
\description{
Processing data frame obtained by \code{\link[=read_spec]{read_spec()}} to be released at a
respective herbarium.

Some herbaria have a template for storage of data.
}
