% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_labels.R
\name{write_labels}
\alias{write_labels}
\alias{write_labels.specimens}
\alias{write_labels,specimens-method}
\title{Writting Labels for Herbarium Vouchers}
\usage{
write_labels(x, ...)

\method{write_labels}{specimens}(
  x,
  output_file,
  merge = TRUE,
  frame = FALSE,
  date_format = "\%d.\%m.\%Y",
  ...
)
}
\arguments{
\item{x}{A \linkS4class{specimens} object retrieved by \code{\link[=read_spec]{read_spec()}}.}

\item{...}{Further arguments passed to \code{\link[=render_rmd]{render_rmd()}}.}

\item{output_file}{A character value indicating the name of the file. It may
include a relative path. This is passed to the homonymous parameter at
\code{\link[=render_rmd]{render_rmd()}}.}

\item{merge}{A logical value indicating whether labels should be merged into
one A4 sheet or remain as individual pages in the output PDF file.}

\item{frame}{A logical value indicating whether a frame should be drawn
around single labels or not. It works only if \code{'merge = TRUE'}.}

\item{date_format}{A character value indicating the format used for the
collection date. It is passed to the function \code{\link[=format.Date]{format.Date()}}.}
}
\value{
An invisible \linkS4class{rmd_doc} object.
}
\description{
Generating labels for vouchers using exported tables.
}
